added delete and update methods with validations

+++ routes/ideas.js
@@ -64,18 +64,28 @@ router.post("/", async (req, res) => {
 router.put("/:id", async (req, res) => {
     try
     {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag,
-                }
-            },
+        const idea = await Idea.findById(req.params.id)
 
-            { new: true }
-        )
-        res.json({ success: true, data: updatedIdea })
+        // match the usernames
+        if (idea.username === req.body.username)
+        {
+
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag,
+                    }
+                },
+
+                { new: true }
+            )
+            return res.json({ success: true, data: updatedIdea })
+        }
+
+        // usernames do not match
+        return res.status(403).json({ success: false, error: "You are not authorized to update this idea" })
 
     } catch (error)
     {
@@ -85,16 +95,23 @@ router.put("/:id", async (req, res) => {
 })
 
 
-
-
-
 // delete idea
 
 router.delete("/:id", async (req, res) => {
     try
     {
-        await Idea.findByIdAndDelete(req.params.id)
-        res.json({ success: true, data: {} })
+        const idea = await Idea.findById(req.params.id)
+
+        // match the usernames
+        if (idea.username === req.body.username)
+        {
+
+            await Idea.findByIdAndDelete(req.params.id)
+            return res.json({ success: true, data: {} })
+        }
+
+        // usernames do not match
+        return res.status(403).json({ success: false, error: "You are not authorized to delete this idea" })
 
     } catch (error)
     {
